Make a library application
As documented https://learnyousomeerlang.com/building-otp-applications

```
What happens when we want to wrap flat modules in an application but we have no process to start and thus no need for an application callback module?

After pulling our hair and crying in rage for a few minutes, the only other thing left to do is to remove the tuple {mod, {Module, Args}} from the application file. That's it. This is called a library application. If you want an example of one, the Erlang stdlib (standard library) application is one of these.
```

Otherwise erlang will try to start the application and it will fail as
eraven doesn't start anything